removed unused routes, moved components into own folder
