+++ src/components/PlantCell.svelte
