+++ src/components/PopupInfo.svelte
